NFCv2: New NFC engagement protocol.
This PR implements NFCv2 according to the latest documents in Action
Point 112.05 in ISO/IEC SC17 WG10.

Also fix a bunch of uses of `java.lang.Error` to throw errors which
are supposed to be recoverable, so use `IllegalStateException`
instead. The problem with using `java.lang.Error` is that it's derived
from `Throwable` - not `Exception` - meaning that it's not catchable
unless you catch `Throwable` which is a big no-no according to
CODING-STYLE.md. Also add language to CODING-STYLE.md to recommend
against throwing this exception.

Go through all relevant code for NFC engagement
(`MdocTransport`-derived classes, `MdocNfcEngagementHelper`,
`MdocNdefService`, and the same for the new V2 variants) and make sure
it's using Kotlin Coroutines correctly. A number of race conditions
and non-ideomatic uses were fixed.

Properly handle when Android re-uses a `HostApduService` for a second
tap soon after a first tap, that is, when `onDestroy` / `onCreate`
isn't called. The symptom was that we would hang on the second
tap. With these changes NFC engagement is now bullet-proof.

This is a work-in-progress so implementation details can and will
change as the small group working on AP 112.05 meets. The scaffolding
however - `MdocNfcV2Service`, `MdocNfcV2EngagementHelper` and
`MdocReaderNfcHandoverOptions` - are expected to not change. This
means that it should be safe to start using this in Multipaz Wallet
and other applications without any API changes.

Test: New unit tests and manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>